refactor: move ats harvest seam below commands into discovered harvest module
drops the l3-to-l3 peer import flagged in review: harvest_ats_refs + posting_to_ref move from
commands/discovery.rs to discovered/harvest.rs, apphandle-free (take the store directly). all 5 call
sites (scrape x3, autopilot, extension import) resolve the store at the shell boundary and pass it
down; missing store stays a graceful no-op. the moved tests now drive the real harvest fn, pinning
the (url, company) arg order the review found unguarded.
